+++ conjur-api/Variable.cs
@@ -68,10 +68,10 @@ namespace Conjur
         /// <param name="client">Conjur client to query.</param>
         /// <param name="query">Query for search.</param>
         /// <returns>Returns IEnumerable to Variable.</returns>
-        internal static IEnumerable<Variable> List(Client client, string query = null)
+        internal static IEnumerable<Variable> List(Client client, string query = null, string actingAs = null)
         {
             Func<ResourceMetadata, Variable> newInst = (searchRes) => new Variable(client, searchRes.Id);
-            return ListResources<Variable, ResourceMetadata>(client, "variable", newInst, query);
+            return ListResources<Variable, ResourceMetadata>(client, "variable", newInst, query, actingAs);
         }
     }
 }