+++ CHANGELOG.md
@@ -5,6 +5,12 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+## 1.3.1 - 2018-05-23
+### Added
+- actungAs parameter to Client.ListVariables and Client.ListUsers methods
+### Removed
+- Client.ActingAs property
+
 ## 1.3.0 - 2018-05-22
 ### Added
 - User entity